+++ src/backend/src/main/java/com/personal/finance/backend/budgets/dto/request/UpdateBudgetRequest.java
@@ -1,16 +1,19 @@
 package com.personal.finance.backend.budgets.dto.request;
 
+import jakarta.persistence.Column;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import lombok.Getter;
 import lombok.Setter;
+
+import java.math.BigDecimal;
 
 @Getter
 @Setter
 public class UpdateBudgetRequest {
     @NotNull(message = "Hạn mức không được để trống")
     @Positive(message = "Hạn mức ngân sách phải lớn hơn 0")
-    private Double limitAmount;
+    private BigDecimal limitAmount;
 
     @Positive(message = "Tỷ lệ cảnh báo phải lớn hơn 0")
     private Double warningPercent;
