+++ src/backend/src/main/java/com/personal/finance/backend/savingGoals/dto/response/SavingGoalDTO.java
@@ -3,6 +3,7 @@ package com.personal.finance.backend.savingGoals.dto.response;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -10,8 +11,8 @@ import java.time.LocalDate;
 public class SavingGoalDTO {
     private Long id;
     private String title;
-    private Double targetAmount;
-    private Double currentAmount;
+    private BigDecimal targetAmount;
+    private BigDecimal currentAmount;
     private LocalDate deadline;
     private String status;
 }
