+++ src/backend/src/main/java/com/personal/finance/backend/transactions/dto/request/CreateTransactionRequest.java
@@ -1,11 +1,13 @@
 package com.personal.finance.backend.transactions.dto.request;
 
 import com.personal.finance.backend.transactions.entity.Transaction;
+import jakarta.persistence.Column;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -19,7 +21,7 @@ public class CreateTransactionRequest {
 
     @NotNull(message = "Số tiền không được để trống")
     @Positive(message = "Số tiền giao dịch phải lớn hơn 0")
-    private Double amount;
+    private BigDecimal amount;
 
     @NotNull(message = "Loại giao dịch không được để trống")
     private Transaction.TransactionType type;
